Don't need to create a new instance of InnerTube when deciphering
