+++ innertube/src/main/java/com/zionhuang/innertube/models/response/PlayerResponse.kt
@@ -1,8 +1,8 @@
 package com.zionhuang.innertube.models.response
 
-import com.zionhuang.innertube.InnerTube
 import com.zionhuang.innertube.models.ResponseContext
 import com.zionhuang.innertube.models.Thumbnails
+import com.zionhuang.innertube.utils.decodeCipher
 import kotlinx.serialization.Serializable
 
 /**
@@ -63,7 +63,7 @@ data class PlayerResponse(
             val isAudio: Boolean
                 get() = width == null
 
-            fun findUrl() = url ?: signatureCipher?.let { InnerTube().decodeCipher(it) }!!
+            fun findUrl() = url ?: signatureCipher?.let { decodeCipher(it) }!!
         }
     }
