+++ innertube/src/main/java/com/zionhuang/innertube/utils/Utils.kt
@@ -2,6 +2,8 @@ package com.zionhuang.innertube.utils
 
 import com.zionhuang.innertube.YouTube
 import com.zionhuang.innertube.pages.PlaylistPage
+import io.ktor.http.URLBuilder
+import io.ktor.http.parseQueryString
 import java.security.MessageDigest
 
 suspend fun Result<PlaylistPage>.completed() = runCatching {
@@ -89,7 +91,7 @@ fun nSigDecode(n: String): String {
         val currentChar = step5[index]
         val indexInCharset =
             (charset.indexOf(currentChar) - charset.indexOf(mutableKeyList[index % mutableKeyList.size]) + index + charset.size - index) %
-                charset.size
+                    charset.size
         transformedChars[index] = charset[indexInCharset]
         mutableKeyList[index % mutableKeyList.size] = transformedChars[index]
     }
@@ -112,4 +114,16 @@ fun sigDecode(input: String): String {
             append(input[38])
         }
     return result
+}
+
+fun decodeCipher(cipher: String): String? {
+    val params = parseQueryString(cipher)
+    val signature = params["s"] ?: return null
+    val signatureParam = params["sp"] ?: return null
+    val url = params["url"]?.let { URLBuilder(it) } ?: return null
+    val n = url.parameters["n"]
+    url.parameters["n"] = nSigDecode(n.toString())
+    url.parameters[signatureParam] = sigDecode(signature)
+    url.parameters["c"] = "ANDROID_MUSIC"
+    return url.toString()
 }
